tools/pyproject.toml
[project]
name = "ontology-tools"
version = "0.1.0"
description = "IT 도메인 학습 문서(DOC/DOCX)를 Neo4j 지식그래프로 변환하는 파이프라인 도구"
requires-python = ">=3.11"

dependencies = [
    "python-docx>=1.1.0",
    "neo4j>=5.20.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "mammoth>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]
win = [
    "pywin32>=306",
]

[project.scripts]
ontology = "cli:main"

[tool.ruff]
line-length = 120
target-version = "py311"
